Add WireMock and integration tests for Tasmota and Shelly Pro
Added WireMock dependency for mocking HTTP interactions in tests. Introduced integration tests for Tasmota and Shelly Pro relay adapters using the new FileOperationUtils and PrefixPath utilities. Configured necessary test data and mappings adjustments.

+++ src/main/java/cz/patyk/solarmaxx/backend/dto/data/RelayOutputDataDto.java
@@ -16,7 +16,7 @@ public class RelayOutputDataDto implements DtoDataInterface {
     private Long relayId;
     private String relayName;
     private String relayIpAddress;
-    private Short relayPort;
+    private int relayPort;
     private Long relayTypeId;
     private String relayTypeName;
     private String relayTypeString;
